Refactor with changes from develop
