+++ src/lib/dataviews/category/CategoryBase.ts
