--- src/lib/widget/index.ts
@@ -1,1 +1,0 @@
-export { Category } from './category/Category';
